DQB-27 Migrate to pydantic and SQLModel
Update pyproject.toml to include SQLModel.
Update docker-compose.yml to continue to better support dynamic local development
