+++ pyproject.toml
@@ -31,7 +31,7 @@ exclude = [
 local_partial_types = true
 namespace_packages = true
 plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
-python_version = "3.10"
+python_version = "3.11"
 show_error_codes = true
 warn_redundant_casts = true
 warn_return_any = true
@@ -66,6 +66,7 @@ psycopg2-binary = "^2.9.6"
 pydantic = "^2.5.3"
 pydantic-settings = "^2.1.0"
 python = "^3.11"
+sqlmodel = "^0.0.14"
 typing-extensions = "^4.9.0"
 
 [tool.poetry.group.dev.dependencies]
@@ -118,7 +119,7 @@ select = [
   "ARG",
   "RUF100"
 ]
-target-version = "py310"
+target-version = "py311"
 
 [tool.ruff.isort]
 force-single-line = true
