+++ src/models.py
@@ -1,13 +1,11 @@
 from datetime import datetime
 from typing import TYPE_CHECKING
 
+from pydantic import ConfigDict
 from sqlalchemy import BigInteger
-from sqlalchemy import ForeignKey
-from sqlalchemy.orm import Mapped
-from sqlalchemy.orm import MappedAsDataclass
 from sqlalchemy.orm import declared_attr
-from sqlalchemy.orm import mapped_column
-from sqlalchemy.orm import relationship
+from sqlmodel import Field
+from sqlmodel import Relationship
 
 from src.helpers.sqlalchemy_helpers import BaseModel
 from src.helpers.sqlalchemy_helpers import snake_case_table_name
@@ -18,20 +16,22 @@ if TYPE_CHECKING:
     from src.quests import Quest
 
 
-class CoreModelMixin(MappedAsDataclass, BaseModel, kw_only=True):
-    @declared_attr  # type: ignore[arg-type]
+class CoreModelMixin(BaseModel):
+    __abstract__ = True
+    model_config = ConfigDict(str_strip_whitespace=True)  # type: ignore[assignment]
+
+    @declared_attr  # type: ignore[override, arg-type]
     def __tablename__(self) -> str:
         return snake_case_table_name(self.__name__)
 
-    __abstract__ = True
-    id: Mapped[int] = mapped_column(init=False, primary_key=True)
-    datetime_created: Mapped[datetime] = mapped_column(init=False, default_factory=datetime.utcnow, repr=False)
-    datetime_edited: Mapped[datetime] = mapped_column(
-        init=False, default_factory=datetime.utcnow, repr=False, onupdate=datetime.utcnow
+    id: int | None = Field(primary_key=True, default=None)
+    datetime_created: datetime = Field(default_factory=datetime.utcnow, repr=False)
+    datetime_edited: datetime = Field(
+        default_factory=datetime.utcnow, repr=False, sa_column_kwargs={"onupdate": datetime.utcnow}
     )
 
 
-class User(CoreModelMixin):
+class User(CoreModelMixin, table=True):
     """
     Representation of a registered discord user
 
@@ -45,23 +45,17 @@ class User(CoreModelMixin):
     """
 
     # Columns
-    discord_id: Mapped[int] = mapped_column(BigInteger, unique=True)
+    discord_id: int = Field(sa_type=BigInteger, unique=True)
 
     # Relationships
-    quests: Mapped[list["Quest"]] = relationship("UserQuest", default_factory=list, repr=False, back_populates="user")
-    experience: Mapped[list["ExperienceTransaction"]] = relationship(
-        "ExperienceTransaction", back_populates="user", default_factory=list, repr=False
-    )
+    quests: list["Quest"] = Relationship(sa_relationship_args=["user_quest"], back_populates="users")
+    experience: list["ExperienceTransaction"] = Relationship(back_populates="user")
 
 
-class UserResourceMixin(MappedAsDataclass):
+class UserResourceMixin(BaseModel):
+    __abstract__ = True
+
     class Meta:
         user_mixin_data: MixinData = MixinData()
 
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("user.id"), init=False, repr=False, index=Meta.user_mixin_data.get("index")
-    )
-
-    @declared_attr
-    def user(self) -> Mapped[User]:
-        return relationship(User, back_populates=self.Meta.user_mixin_data.get("back_populates"))
+    user_id: int = Field(foreign_key="user.id", repr=False, index=Meta.user_mixin_data.index)
