+++ src/tavern/models.py
@@ -1,20 +1,17 @@
 from datetime import date
 from datetime import datetime
 
-from sqlalchemy import ForeignKey
-from sqlalchemy.orm import Mapped
-from sqlalchemy.orm import mapped_column
-from sqlalchemy.orm import relationship
+from sqlmodel import Field
+from sqlmodel import Relationship
 
 from src.constants import DayOfWeek
-from src.helpers.sqlalchemy_helpers import EnumColumn
 from src.models import CoreModelMixin
 
 
-class Menu(CoreModelMixin):
-    server_id: Mapped[int]
-    start_date: Mapped[date] = mapped_column(default_factory=datetime.today)
-    items: Mapped[list["MenuItem"]] = relationship("MenuItem", back_populates="menu", default_factory=list)
+class Menu(CoreModelMixin, table=True):
+    server_id: int
+    start_date: date = Field(default_factory=datetime.today)
+    items: list["MenuItem"] = Relationship(back_populates="menu")
 
     @property
     def grouped_items(self) -> dict[DayOfWeek, list["MenuItem"]]:
@@ -24,12 +21,12 @@ class Menu(CoreModelMixin):
         return dow_items
 
 
-class MenuItem(CoreModelMixin):
+class MenuItem(CoreModelMixin, table=True):
     # Columns
-    food: Mapped[str]
-    day_of_the_week: Mapped[DayOfWeek] = mapped_column(EnumColumn(DayOfWeek))
+    food: str
+    day_of_the_week: DayOfWeek = Field()
 
-    menu_id: Mapped[int] = mapped_column(ForeignKey("menu.id", ondelete="CASCADE"), init=False, repr=False)
+    menu_id: int = Field(foreign_key="menu.id", repr=False)
 
     # relationships
-    menu: Mapped[Menu] = relationship(Menu, back_populates="items", init=False)
+    menu: Menu = Relationship(back_populates="items")
